requirements.txt
click
gitpython
loguru
platformdirs
pyyaml
